Impl. and use `env_var.hpp`
